Clases y demases

+++ dominio/Piton.cs
@@ -11,7 +11,7 @@ namespace dominio
         public int Id { get; set; }
         public string Modelo { get; set; }
         public int Cantidad { get; set; }
-        public float Peso { get; set; }
+        public double Peso { get; set; }
 
         public override string ToString()
         {
